Use smoot components; fix radio button focus ring (#2304)
* use checkboxes and radio buttons from smoot

* remove ol-components VisuallyHidden

* remove ol-components Alert

* remove TextField

* remove Input from ol-components

* use Tabs from smoot-design
