+++ frontends/main/src/app-pages/ChannelPage/ChannelSearch.tsx
@@ -5,7 +5,8 @@ import { useResourceSearchParams } from "@mitodl/course-search-utils"
 import type { Facets, BooleanFacets } from "@mitodl/course-search-utils"
 import { useSearchParams } from "@mitodl/course-search-utils/next"
 import SearchDisplay from "@/page-components/SearchDisplay/SearchDisplay"
-import { Container, styled, VisuallyHidden } from "ol-components"
+import { Container, styled } from "ol-components"
+import { VisuallyHidden } from "@mitodl/smoot-design"
 import { SearchField } from "@/page-components/SearchField/SearchField"
 import { getFacets } from "./searchRequests"
 import { keyBy } from "lodash"
