+++ frontends/main/src/app-pages/DashboardPage/CoursewareDisplay/DashboardDialogs.tsx
@@ -5,10 +5,8 @@ import {
   FormDialog,
   DialogActions,
   Stack,
-  Alert,
-  Checkbox,
 } from "ol-components"
-import { Button } from "@mitodl/smoot-design"
+import { Button, Checkbox, Alert } from "@mitodl/smoot-design"
 
 import NiceModal, { muiDialogV5 } from "@ebay/nice-modal-react"
 import { useFormik } from "formik"
