+++ frontends/main/src/app-pages/SearchPage/SearchPage.tsx
@@ -6,7 +6,8 @@ import type { FacetManifest } from "@mitodl/course-search-utils"
 import { useSearchParams } from "@mitodl/course-search-utils/next"
 import { useResourceSearchParams } from "@mitodl/course-search-utils"
 import SearchDisplay from "@/page-components/SearchDisplay/SearchDisplay"
-import { styled, Container, theme, VisuallyHidden } from "ol-components"
+import { styled, Container, theme } from "ol-components"
+import { VisuallyHidden } from "@mitodl/smoot-design"
 import { SearchField } from "@/page-components/SearchField/SearchField"
 import { useOfferorsList } from "api/hooks/learningResources"
 import { facetNames } from "./searchRequests"
