+++ frontends/main/src/page-components/ManageListDialogs/ManageListDialogs.tsx
@@ -2,16 +2,13 @@ import React, { useCallback } from "react"
 import { useFormik, FormikConfig } from "formik"
 import NiceModal, { muiDialogV5 } from "@ebay/nice-modal-react"
 import { RiDeleteBinLine } from "@remixicon/react"
+import { Autocomplete, FormDialog, Dialog, MenuItem } from "ol-components"
 import {
+  Button,
+  BooleanRadioChoiceField,
   Alert,
   TextField,
-  Autocomplete,
-  BooleanRadioChoiceField,
-  FormDialog,
-  Dialog,
-  MenuItem,
-} from "ol-components"
-import { Button } from "@mitodl/smoot-design"
+} from "@mitodl/smoot-design"
 import * as Yup from "yup"
 import { PrivacyLevelEnum, type LearningPathResource, UserList } from "api"
