+++ frontends/main/src/page-components/SearchDisplay/SearchDisplay.tsx
@@ -10,11 +10,14 @@ import {
   truncateText,
   css,
   Drawer,
-  Checkbox,
-  VisuallyHidden,
   Stack,
 } from "ol-components"
-import { Button, ButtonProps } from "@mitodl/smoot-design"
+import {
+  Button,
+  ButtonProps,
+  childCheckboxStyles,
+  VisuallyHidden,
+} from "@mitodl/smoot-design"
 
 import {
   RiCloseLine,
@@ -195,7 +198,7 @@ const FacetStyles = styled.div`
         cursor: pointer;
       }
 
-      ${Checkbox.styles}
+      ${({ theme }) => childCheckboxStyles(theme)}
 
       .facet-count {
         font-size: 12px;
