+++ frontends/ol-components/src/index.ts
@@ -64,12 +64,6 @@ export type { TabProps } from "@mui/material/Tab"
 export { default as TabList } from "@mui/lab/TabList"
 export type { TabListProps } from "@mui/lab/TabList"
 
-export {
-  TabButton,
-  TabButtonLink,
-  TabButtonList,
-} from "./components/TabButtons/TabButtonList"
-
 export { default as TabContext } from "@mui/lab/TabContext"
 export type { TabContextProps } from "@mui/lab/TabContext"
 export { default as TabPanel } from "@mui/lab/TabPanel"
@@ -102,7 +96,6 @@ export type { StepIconProps } from "@mui/material/StepIcon"
 export { default as FormGroup } from "@mui/material/FormGroup"
 export { default as Slider } from "@mui/material/Slider"
 
-export * from "./components/Alert/Alert"
 export * from "./components/BannerPage/BannerPage"
 export * from "./components/Breadcrumbs/Breadcrumbs"
 export * from "./components/Card/Card"
@@ -110,8 +103,6 @@ export * from "./components/Card/ListCardCondensed"
 export * from "./components/Carousel/Carousel"
 export { onReInitSlickA11y } from "./components/Carousel/util"
 
-export * from "./components/Checkbox/Checkbox"
-export * from "./components/Checkbox/CheckboxChoiceField"
 export * from "./components/Chips/ChipLink"
 export * from "./components/ChoiceBox/ChoiceBox"
 export * from "./components/ChoiceBox/ChoiceBoxField"
@@ -132,20 +123,15 @@ export * from "./components/SimpleMenu/SimpleMenu"
 export * from "./components/SortableList/SortableList"
 export * from "./components/ThemeProvider/ThemeProvider"
 export * from "./components/TruncateText/TruncateText"
-export * from "./components/Radio/Radio"
-export * from "./components/RadioChoiceField/RadioChoiceField"
-export * from "./components/VisuallyHidden/VisuallyHidden"
 
 export * from "./constants/imgConfigs"
 
-export { Input, AdornmentButton } from "./components/Input/Input"
-export type { InputProps, AdornmentButtonProps } from "./components/Input/Input"
 export { SearchInput } from "./components/SearchInput/SearchInput"
 export type {
   SearchInputProps,
   SearchSubmissionEvent,
 } from "./components/SearchInput/SearchInput"
-export { TextField } from "./components/TextField/TextField"
+
 export {
   SimpleSelect,
   SimpleSelectField,
@@ -156,7 +142,6 @@ export type {
   SimpleSelectOption,
 } from "./components/SimpleSelect/SimpleSelect"
 
-export type { TextFieldProps } from "./components/TextField/TextField"
 export { SelectField } from "./components/SelectField/SelectField"
 export type {
   SelectChangeEvent,
